Moves `RemoteFeedImageDataLoader` to production
